added exceptions for libraryservice and fixed module settings

+++ LibraryService/src/main/java/com/example/libraryservice/controller/FolderController.java
@@ -1,13 +1,15 @@
-package src.main.java.com.example.libraryservice.controller;
+package com.example.libraryservice.controller;
 
+import com.example.libraryservice.errors.FolderNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
 
-import src.main.java.com.example.libraryservice.service.FolderService;
-import src.main.java.com.example.libraryservice.entity.Folder;
+import com.example.libraryservice.model.FolderRequestModel;
+import com.example.libraryservice.service.FolderService;
+import com.example.libraryservice.entity.Folder;
 
 @RestController
 @RequestMapping("/folders")
@@ -17,33 +19,35 @@ public class FolderController {
     private FolderService folderService;
 
     @GetMapping("/{id}")
-    public Folder FindFolderById(@PathVariable("id") Long folderId){
-//        if (folderService.findFolderById(folderId) == null){
-//            throw new com.example.fileservice.errors.LocationNotFoundException("Location not found");
-//        }
+    public Folder findFolderById(@PathVariable("id") Long folderId){
+        if (folderService.findByFolderId(folderId) == null){
+            throw new FolderNotFoundException("Folder not found");
+        }
         return folderService.findByFolderId(folderId);
     }
 
     @GetMapping("/")
-    public List<Folder> GetAllFolders() {
+    public List<Folder> getAllFolders() {
         return folderService.getAllFolders();
     }
 
     @PostMapping("/add")
-    public Folder saveLocation(@Valid @RequestBody Folder newFolder) {
-        return folderService.saveFolder(newFolder);
+    public Folder saveLocation(@Valid @RequestBody FolderRequestModel newFolder) {
+        Folder folder = new Folder(newFolder);
+        return folderService.saveFolder(folder);
     }
 
     @PutMapping("/edit/{id}")
-    public Folder updateLocation(@Valid @RequestBody Folder updatedFolder, @PathVariable Long id) {
-        return folderService.updateFolder(updatedFolder, id);
+    public Folder updateLocation(@Valid @RequestBody FolderRequestModel  updatedFolder, @PathVariable Long id) {
+        Folder folder = new Folder(updatedFolder);
+        return folderService.updateFolder(folder, id);
     }
 
     @DeleteMapping("/delete/{id}")
     public void deleteLocation(@PathVariable Long id) {
-//        if (folderService.findFolderById(id) == null){
-//            throw new com.example.fileservice.errors.LocationNotFoundException("Location not found");
-//        }
+        if (folderService.findByFolderId(id) == null){
+            throw new FolderNotFoundException("Folder not found");
+        }
 
         folderService.deleteById(id);
     }
